new layers.py file
